python/requirements.txt
PyYAML==3.12
rivescript==1.14.2
six==1.10.0
